+++ lib/runners/processor/querly.rb
@@ -3,7 +3,11 @@ module Runners
     include Ruby
 
     Schema = StrongJSON.new do
-      let :runner_config, Schema::BaseConfig.ruby
+      let :runner_config, Schema::BaseConfig.ruby.update_fields { |fields|
+        fields.merge!({
+          config: string?,
+        })
+      }
 
       let :rule, object(
         id: string,
@@ -24,27 +28,16 @@ module Runners
       "querly" => [">= 0.5.0", "< 2.0.0"]
     }.freeze
 
+    CONFIG_FILE = "querly.yml".freeze
+    CONFIG_FILES_GLOB = "querly.{yml,yaml}".freeze
+
     def analyzer_version
       @analyzer_version ||= extract_version! ruby_analyzer_bin, "version"
     end
 
     def setup
-      ret = install_gems default_gem_specs, optionals: OPTIONAL_GEMS, constraints: CONSTRAINTS do
+      install_gems default_gem_specs, optionals: OPTIONAL_GEMS, constraints: CONSTRAINTS do
         yield
-      end
-
-      # NOTE: Exceptionally MissingFileFailure is treated as successful
-      if ret.instance_of? Results::MissingFilesFailure
-        add_warning(<<~MESSAGE)
-          Sider cannot find the required configuration file `querly.yml`.
-          Please set up Querly by following the instructions, or you can disable it in the repository settings.
-
-          - https://github.com/soutaro/querly
-          - #{analyzer_doc}
-        MESSAGE
-        Results::Success.new(guid: guid, analyzer: analyzer)
-      else
-        ret
       end
     rescue InstallGemsFailure => exn
       trace_writer.error exn.message
@@ -52,46 +45,87 @@ module Runners
     end
 
     def analyze(changes)
-      capture3!(*ruby_analyzer_bin, "version")
+      if !config_file && !default_config_file
+        return missing_config_file_result
+      end
 
-      ensure_files relative_path("querly.yml"), relative_path("querly.yaml") do |config_file|
-        test_config_file(config_file)
+      test_config_file
 
-        Results::Success.new(guid: guid, analyzer: analyzer).tap do |result|
-          stdout, _ = capture3!(*ruby_analyzer_bin, "check", "--format=json", ".")
+      stdout, _ = capture3!(
+        *ruby_analyzer_bin,
+        "check",
+        "--format=json",
+        *(config_file ? ["--config", config_file] : []),
+        ".",
+      )
 
-          json = JSON.parse(stdout, symbolize_names: true)
-          Array(json[:issues]).each do |hash|
-            path = relative_path(hash[:script])
-            loc = Location.new(start_line: hash[:location][:start][0],
-                                            start_column: hash[:location][:start][1],
-                                            end_line: hash[:location][:end][0],
-                                            end_column: hash[:location][:end][1])
-            msgs = hash[:rule][:messages]
-            result.add_issue Issue.new(
-              path: path,
-              location: loc,
-              id: hash[:rule][:id],
-              message: (msgs && msgs[0]) || "No message",
-              object: hash[:rule],
-              schema: Schema.rule
-            )
-          end
+      json = JSON.parse(stdout, symbolize_names: true)
+
+      Results::Success.new(guid: guid, analyzer: analyzer).tap do |result|
+        Array(json[:issues]).each do |hash|
+          start_line, start_column = hash[:location][:start]
+          end_line, end_column = hash[:location][:end]
+          rule = hash[:rule]
+          message, _ = rule[:messages]
+
+          result.add_issue Issue.new(
+            path: relative_path(hash[:script]),
+            location: Location.new(
+              start_line: start_line,
+              start_column: start_column,
+              end_line: end_line,
+              end_column: end_column,
+            ),
+            id: rule[:id],
+            message: message || "No message",
+            object: rule,
+            schema: Schema.rule
+          )
         end
       end
     end
 
-    def test_config_file(config_file)
-      stdout, _stderr, _status = capture3(*ruby_analyzer_bin, 'test')
+    private
 
-      warnings = stdout.each_line.map do |line|
-        match = line.match(/\A  (\S+:\t.+)\Z/)
-        next unless match
-        match[1]
-      end.compact
+    def config_file
+      config_linter[:config]
+    end
 
-      warnings.each do |warn|
-        add_warning(warn, file: config_file.to_s)
+    def default_config_file
+      return @default_config_file if defined? @default_config_file
+
+      config_files = Dir.glob(CONFIG_FILES_GLOB)
+
+      if config_files.size > 1
+        file_list = config_files.map { |file| "`#{file}`" }.join(', ')
+        add_warning <<~MSG, file: CONFIG_FILE
+          There are duplicate configuration files (#{file_list}). Remove the files except the first one.
+        MSG
+      end
+
+      @default_config_file ||= config_files.first
+    end
+
+    def missing_config_file_result
+      add_warning <<~MSG, file: CONFIG_FILE
+        Sider could not find the required configuration file `#{CONFIG_FILE}`.
+        Please create the file according to the following documents:
+        - #{analyzer_github}
+        - #{analyzer_doc}
+      MSG
+
+      Results::Success.new(guid: guid, analyzer: analyzer)
+    end
+
+    def test_config_file
+      stdout, _stderr, _status = capture3(
+        *ruby_analyzer_bin,
+        "test",
+        *(config_file ? ["--config", config_file] : []),
+      )
+
+      stdout.scan(/^  (\S+:\t.+)$/) do |message, _|
+        add_warning message, file: config_file || default_config_file
       end
     end
   end